+++ src/components/metadataFormControls/ModelTransfer/ModelTransferField.tsx
@@ -1,12 +1,9 @@
-import i18n from '@dhis2/d2-i18n'
-import { Button, ButtonStrip, Field, TransferProps } from '@dhis2/ui'
-import React, { useRef } from 'react'
+import { Field, TransferProps } from '@dhis2/ui'
+import React from 'react'
 import { useField } from 'react-final-form'
-import { useHref } from 'react-router'
-import { DisplayableModel, ModelTransfer } from '../../../components'
-import { getSectionNewPath } from '../../../lib'
+import { ModelTransfer } from '../../../components'
 import { PlainResourceQuery } from '../../../types'
-import { LinkButton } from '../../LinkButton'
+import { DisplayableModel } from '../../../types/models'
 import css from './ModelTransfer.module.css'
 
 // this currently does not need a generic, because the value of the field is not passed
@@ -30,24 +27,16 @@ export function ModelTransferField({
     name,
     query,
     label,
-    rightHeader,
     leftHeader,
-    rightFooter,
+    rightHeader,
     leftFooter,
+    rightFooter,
     filterPlaceholder,
     filterPlaceholderPicked,
 }: ModelTransferFieldProps) {
-    const modelName = query.resource
     const { input, meta } = useField<DisplayableModel[]>(name, {
         multiple: true,
         validateFields: [],
-    })
-    const newLink = useHref(`/${getSectionNewPath(modelName)}`)
-
-    const modelTransferHandle = useRef({
-        refetch: () => {
-            throw new Error('Not initialized')
-        },
     })
 
     return (
@@ -60,62 +49,19 @@ export function ModelTransferField({
             className={css.moduleTransferField}
         >
             <ModelTransfer
-                ref={modelTransferHandle}
-                filterPlaceholder={
-                    filterPlaceholder || i18n.t('Filter available items')
-                }
-                filterPlaceholderPicked={
-                    filterPlaceholderPicked || i18n.t('Filter selected items')
-                }
                 selected={input.value}
                 onChange={({ selected }) => {
                     input.onChange(selected)
                     input.onBlur()
                 }}
-                leftHeader={<TransferHeader>{leftHeader}</TransferHeader>}
-                rightHeader={<TransferHeader>{rightHeader}</TransferHeader>}
-                leftFooter={
-                    leftFooter ?? (
-                        <DefaultTransferLeftFooter
-                            onRefreshClick={modelTransferHandle.current.refetch}
-                            newLink={newLink}
-                        />
-                    )
-                }
+                leftHeader={leftHeader}
+                rightHeader={rightHeader}
+                leftFooter={leftFooter}
                 rightFooter={rightFooter}
+                filterPlaceholder={filterPlaceholder}
+                filterPlaceholderPicked={filterPlaceholderPicked}
                 query={query}
-                height={'350px'}
-                optionsWidth="500px"
-                selectedWidth="500px"
-                enableOrderChange={true}
             />
         </Field>
-    )
-}
-
-const TransferHeader = ({ children }: { children: React.ReactNode }) => {
-    if (typeof children === 'string') {
-        return <div className={css.modelTransferHeader}>{children}</div>
-    }
-    return <>{children}</>
-}
-
-const DefaultTransferLeftFooter = ({
-    onRefreshClick,
-    newLink,
-}: {
-    onRefreshClick: () => void
-    newLink: string
-}) => {
-    return (
-        <ButtonStrip className={css.modelTransferFooter}>
-            <Button small onClick={onRefreshClick}>
-                {i18n.t('Refresh list')}
-            </Button>
-
-            <LinkButton small href={newLink} target="_blank">
-                {i18n.t('Add new')}
-            </LinkButton>
-        </ButtonStrip>
     )
 }